2ndProject/Voxalyze/requirements.txt
aiofiles==24.1.0
Authlib==1.3.2
bcrypt==4.2.0
cachetools==5.5.0
dnspython==2.6.1
google-auth==2.34.0
html5tagger==1.3.0
httptools==0.6.1
Jinja2==3.1.3
motor==3.5.1
multidict==6.0.5
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydub==0.25.1
PyJWT==2.9.0
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sanic==24.6.0
Sanic-Cors==2.2.0
sanic-jwt==1.8.0
sanic-routing==23.12.0
tracerite==1.1.1
ujson==5.10.0
uvloop==0.19.0
websockets==12.0
